TelegramBot/requirements.txt
charset-normalizer==2.0.5
email-validator==1.1.3
jurigged==0.3.4
phonenumbers==8.12.31
pydantic==1.8.2
PyJWT==2.1.0
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
typing==3.7.4.3
typing-extensions==3.10.0.2
urllib3==1.26.6
